deprecate trainerproperties mixin and move properties to trainer.py
